refactor: put an error message when a translation that does not exist is passed to it

+++ crates/lang/build.rs
@@ -2,7 +2,7 @@
 mod dialog;
 
 use std::env;
-use std::fs::{read_to_string, File};
+use std::fs::{read_to_string, File, read_dir};
 use std::path::Path;
 use std::io::Write;
 use miniserde::json;
@@ -28,15 +28,30 @@ fn read_struct() -> String {
     output
 }
 
-
-
 fn main() {
     let lang = env::var("LANG".to_string()).unwrap();
-    
-    let j = read_to_string(format!("defs/{lang}.json")).unwrap();
-    let dialogs: Dialogs = json::from_str(&j).unwrap();
- 
+     
     let dest_path = Path::new("src").join("lang_info.rs");
+    let paths: Vec<_> = read_dir("defs").unwrap()
+        .map(|entry| entry.unwrap().path())
+        .collect();
+    let lang_path = Path::new("defs").join(&lang);
+
+    if !lang_path.exists() {
+        eprintln!("A translation called {} was not found", lang);
+        eprintln!("consider using some of these translations");
+        for path in &paths {
+            if path.is_file() {
+                eprintln!("{}", path.display());
+            }
+        }
+        panic!();
+    }
+
+    let j = read_to_string(lang_path).unwrap();
+    let dialogs: Dialogs = json::from_str(&j).unwrap();
+
+
     let mut file = File::create(dest_path).unwrap();
     let def = read_struct();
 
@@ -45,10 +60,7 @@ fn main() {
         pub const DIALOGS: Dialogs = {:?};
     "#, dialogs).as_bytes()).unwrap();
 
-    let paths = std::fs::read_dir("defs").unwrap();
-
-    for path in paths {
-        let path = path.unwrap().path();
+    for path in &paths {
         if path.is_file() {
             println!("cargo:rerun-if-changed={}", path.display());
         }
